Factor out factory_register definitions
Method definitions moved below declaration of factory_register,
with extra template info added to allow this move. No change in
functionality is made. This section can later be refactored out
into a -inl.h file.

This gives a cleaner public interface to factory register.
Some compromises in formatting are made in the definition section
where return types are nested, templated types of factory_register.
Here we break convention and put the (very long) return type on 
the one line and the function name and parameters on the next.
This seems to be the clearest and most readable format.