C++ Standard compliance fixes
Use isinf function from the std namespace inside datatools::is_infinite
for compatibility on Mac. Long term, this needs a general check as it 
is known that the Intel compiler is further different in where isinf
is located and called.

Remove `typename` keyword in check_serial_tag argument. This is needed
for C++03 compatibility.

Tidy up some formatting in top level CMakeLists.txt.