Add priority based iostream logging
Implement datatools::logger struct which holds an enum of priority levels
for log messages. Eight levels are available from PRIO_FATAL, the
highest priority, to PRIO_TRACE, the lowest. Implement a static function
to convert between a string name and the priority enum value. This allows
configurable objects to take a string, taken from for example a
configuration file, as an argument to set their logging level.

Provide macros to automate the logging of messages provided a given
minimum priority is met. These take a priority and message, which may be
constructed using stream formatting.

Add basic doxygen markup and example code.

Add unit tests that simply check the logging macros compile and function,
plus checking of the string to enum conversion.