Simplify priority-string conversion
Logging priorities are mapped to three different valid strings. The
canonical string form for priority PRIO_NAME is "name", and this may
not be found easily from the priority-string lookup table because of
the string sorting.

Provide GetCanonicalLabel function which converts any of the three
valid string priorities to the canonical form. Replace direct 
iterator manipulation in get_priority_label with a call to 
GetCanonicalLabel if the priority is located.

Update test_logger to automatically report errors and fail via
exception throwing when testing get_priority_label.