Major change of the interface
The former  technique to hold user  data within cuts  was using ``void
*`` pointers. This new version  now makes use of dedicated templatized
data wrappers which  enable some explicit testing of  the type of user
data referenced by the cuts. The wrapper concept could also be used to
implement  special  traits/functionalities in  the  future.  The  user
interface  based  on the  use  of a  cut  manager  object and  dynamic
instantiation of cut objects  through factories associted to container
of properties is unchanged.