Add api header for handling deprecation
As APIs like datatools involve, it's important to inform clients of
interface changes as soon as possible. A good way to do this is use
compiler attributes to mark deprecated interfaces so that a warning is
emitted if the client uses these.

Add detail header api.h to datatools, and implement a preprocessor
definition DATATOOLS_DEPRECATED(Message). Define it for Clang, Intel,
Microsoft and GNU compilers to the appropriate __attribute__ or __declspec
marker. Always define it with a message, adjusting the underlying
definition to handle known cases where the compiler supports a
message attached to the deprecation. Otherwise discard the message.

Add new Handling API Changes guide to the doxygen Related Pages.
Write up a rough and ready example of using the header and definition.
TODO: Add information of using Doxygen's deprecated command and
further information on handling changes.