Add SVN support CMake module
Add simple CMake module to support checking and working with
Subversion at configure time. Use CMake's FindSubversion module
for most of the work. Provide simple CMake function to check whether
a passed directory is a Subversion working copy. 

Prepend the cmake/ custom module path to front of CMAKE_MODULE_PATH
and include the module as part of the core CMake setup for Bayeux.