Provide interface for querying build number
Use SVN CMake functions to extract current revision number of
Bayeux and export it as a preprocessor definition in Bayeux's
version header.

Implement version::get_revision() method to return this number.

TODO: Think about how to extract diff/status info, and to export
the revision number in a source package as the number is only currently
generated in a subversion working copy.