Add unit testing
Provide CMake option that when ON, sets up CTest based testing
for Bayeux.

List test sources and environment in the module file for datatools.
Remove reflection and interactive tests for now.

For each test, build a testing executable, linking it to bayeux.
Add a CTest test for each one, prefixing each test name with the module
name so that groups of tests can be run.