Ensure fft_real::main returns
Clang compiler reveals a "control reaches end of non-void function"
warning for fft_real::main static method.

Following return expected for "main" functions, return 0 at end of this
function. Not completely clear if this is correct, but function does
not return anything, and it is assumed that reaching the end of the
function implies successful completion.