Integrate brio submodule
Add module file describing brio headers, sources and tests. Use
Cadfael's root_generate_dictionary macro to create required ROOT
dictionary for `TArrayCMod` and `brio_record` classes. Implement
a linkdef file for these classes to allow override of streamer
generation of TArrayCMod.

Add find_package for ROOT in main CMake script of source/, and link
Bayeux to the ROOT libraries. This results in some overlinking to
ROOT, but fixing this requires updates to Cadfael.

This commit works with Cadfael at the current trunk, and does not as
yet completely allow usage of non-Cadfael versions of the required
third-party packages.