Disable shift-count-overflow Clang warning
Building datatools/eos/portable_oarchive.hpp with the Clang compiler
results in a shift-count-overflow warning due to the statement:

do { temp >>= CHAR_BIT; ++size; } 
        
at line 326 of this commit. This appears benign, but should be checked.

Because the line appears in the header, and hence in many downstream files,
wrap the statement in clang pragmas to disable this warning at the location
described above.