Provide embeddable build option
Bayeux may be embedded as an external package - canonical use case
is inside the Falaise bundle.

Provide Bayeux_EMBEDDED variable which parent projects should set
to inform Bayeux that it is being used as an embedded project.
Allow Bayeux_BUILDPRODUCT_DIR to be set by the parent project so
that published headers and resources may be easily integrated.
This technique is primarily to help integrate documentation.