Integrate MCTools Geant4 module
Add list of headers/sources of mctools' Geant4 implementation to module
file for bxmctools. List them in separate variables to separate them
from the main Bayeux sources as we don't want that to depend on Geant4.
Do not export mctools/g4 headers for now as it is not clear if these
are really required.

Wrap creation of library inside the Bayeux_WITH_GEANT4 option, finding
Geant4 with gdml support as part of this. Create library as a MODULE type
so that it need not be directly linked to.

TODO : Find out which headers are really needed. Appears only manager
and manager_parameters appear in actual applications.