Add ROOT Core library to link list
Reported by François Mauger, linking of executables against Bayeux
with gcc/ld can fail with errors about missing symbols that the linker
reports as being in ROOT's libCode. Triaged this as due to usage of
the --no-copy-dt-needed-entries and --as-needed arguments on Ubuntu
systems - reproduced the error on SUSE by adding these flags to the
linker arguments.

Add Core component to list of ROOT components required by Bayeux, and
explicitly link libBayeux to it.