Integrate and test MCTools Geant4 module
Move Bayeux_WITH_GEANT4 option to top level CMake script to make it
visible to both source and examples.

Extend MCTools module file to list Geant4 extension module headers and
sources. Only publish public headers, listing detail headers together
with the sources to ensure dependency pickup. Build separate module
library for the extension module, linking it only to the Geant4
libraries to emphasize its nature as an extension module. Add
commands to install the module in the same location as the main
Bayeux library.

Export the Geant4 extension module to the BayeuxConfig.cmake module
file, setting _FOUND style variable to indicate whether it is present.
As the module is expected to hide the use of Geant4 behind the module
interface, do not export re-finding of Geant4 as is done for other Bayeux
dependencies.

Add new example/test bxmctools_geant4 with ExternalProject support to
test location and use of Geant4 extension module. It only checks that
the manager/manager_params objects can be instantiated, and does
not run an active simulation.