Provide infrastructure for exporting Doxygen
Clients of Bayeux may with to use/point to the Doxygen documentation
of Bayeux for help or for cross-referencing.

Provide a cmake file exporting the Doxygen HTML output directory path
plus the Doxygen tag file. Install this file alongside main BayeuxConfig
cmake file. Include the Doxygen cmake file in BayeuxConfig, protecting
it with an OPTIONAL argument to handle case when documentation was not
built.

TODO : Export correct relative paths so HTML/Tagfile may be located
relocatably.