Fix Mac OS X RPATHs
Update linking of Bayeux binaries to use correct rpath options and
commands on Mac OS X. Set library install_name_dirs to @rpath so that
they are relocatable, at the cost of requiring client executables to
set their rpath correctly. Set executable rpaths to @loader_path plus
the relative path from the executable to library directory. Add
the required link flags to the executable to activate this on CMake
versions prior to 2.8.12 

See these article/posts for information on RPATH handling in CMake:

- http://www.kitware.com/blog/home/post/510
- http://www.mail-archive.com/cmake@cmake.org/msg47143.html
- http://www.cmake.org/Wiki/CMake_RPATH_handling