add support for metadata in I/O module (to be validated)
I/O modules (namely dpp::input_module and dpp::output_module classes)
are responsible of the event record (datatools::things instances)
load/store operations from/in files (XML, ASCII, binary or brio format).

The current release now enables the serialization of some metadata
in the filestream, whatever the underlying format is). The metadata are
instantiated as an array of arbitrary datatools::properties objects. They
are stored before the stream of event records. A dedicated container within
the [input/output]_module classes, the "metadata store"
datatools::multi_properties object, can be used to manipulate the various
metadata objects that comes with the event records.

An example program is given in examples/ex02.

This concept may be used to implement some kind of "run header" metadata
structure that would be stored before a set of event records:

  <archive metadata="run_header" type="datatools::properties"/>
  <archive data="event_record"     type="datatools::things"/>
  <archive data="event_record"     type="datatools::things"/>
  <archive data="event_record"     type="datatools::things"/>
  ...
  <archive data="event_record"     type="datatools::things"/>