Re-implement and extend metadata support in I/O modules.
The 'dpp::input_module' and  'dpp::output_modules' classes support two
different metadata systems:

 * Embedded metadata: this is an  internal container to the I/O module
   (a 'datatools::multi_properties' object) in which it is possible to
   store  various arbitrary  parameters  (the metadata  are stored  in
   'sections' as 'datatools::properties' objects).

 * External metadata from a  'dpp::context_service object': this is an
   external container  (also a  'datatools::multi_properties') managed
   by  a dedicated  service which  is  independant of  the I/O  module
   lifetime cycle and  thus can be re-used by other  modules along the
   data processing pipeline.  Some (or  all) sections of this external
   store can  be used as  metadata, serialized together with  the data
   record.

Both systems can coexists with I/O modules. So it is possible to store
in the same output file selected metadata from the context service and
metadata from the embedded store. It is guaranted that in a given data
file, all  requested metadata records  are serialized before  the data
records   (first  external,   then  embedded).

The  deserialization mechanism  now  implies the  stored metadata  are
loaded before  the first data  record from an  input file.  This  is a
different behaviour than  the original implementation in  which it was
needed to  load at least  one event to  guarantee the loading  of some
heading metadata.

The implementation may be difficult to  use for the special case of an
I/O module  handling a list of  files. It is in  principle possible to
modify the  content and values  of the metadata between  two different
sets of data records stored in  two different files, each with its own
metadata  set.   While using  a  'dpp::input_module'  object, if  some
client  application  relies  on  the  metadata  to  setup  some  other
resources, it  may be  necessary to manually  synchronize/update these
resources with the newly loaded metadata when a new file is open along
the   pipeline,    possibly   with   a   new    metadata   set.    The
'dpp::input_module::metadata_was_updated' method is  provided to alert
when such a  case may occur. It  is the responsibility of  the user to
implement his/her  own I/O  system aware  of 'metadata  change'. There
could be some issues in the design of some data record processing loop
due to  the ordering  of the  'load data  record' and  'check metadata
update flag' operations.  The current implementation could be reviewed
in the future if it turns to create problems.