fix tolerance issue while computing phase space factor in Majoron M1 DBD mode"
The calculation of the electron differential phase space (energy spectrum)
implies to integrate some term using the Gauss Kronrod method from the GSL.
In the special case of computing this P.D.F. near the Qbb end-point, the
default requested tolerance (relative error = 1e-4) is too small
and cannot be satisfied by the highest order intergration loop.

Solution: when the tolerance is not reached by the integration method,
another pass is done with a least-demanding tolerance (max: 3 attempts).

Also the GSL error is now intercepted in order to process with our own
error handling.