Check both fail/badbit and eof of istreams
With Clang (Apple) on Mavericks, it appears that the fail/badbits
of an istringstream are set when eof is reached. Current research
suggests this is o.k. behaviour with C++98, but C++11 specifies that
eof should not result in error bits being set.

The parsing of tokens in the id_selector class can result in the
parsed istringstream being checked for failure when it is at end of
file, resulting in an exception being thrown for an otherwise correct
parse. Update check to test both failure bits and eof state. Only
throw exception if stream is bad and not at eof.