Fix incorrect logic when setting internal seed
Use of internal random number generator results in an exception when
setting the internal seed. Tracked to reversed logic in 
set_embeded_prng_seed method. Intent is to throw when instance is
initialized, but logic results in throw when not initialized.

Fix logic so that throw happens only when instance is already initialized.