Refactor the gnuplot color display support
Gnuplot --  through its 'plot' command  -- can use a  "variable color"
field to  color the  graphics objects  it displays  (points, lines...)
with  respect to  a given  color map.   Here we  have implemented  the
support for some "color::context" object attached to the generation of
the data to be  displayed using primitive functions (gnuplot_draw...).
This enables  the generation of (x,  y, z) triplets associated  with a
color index.  This 'color context' can  be activated on demand and the
user is free to use this field.

Also the 'bounding box' support have been refactored.  Both mechanisms
use  a singleton  to  let the  user interface  as  simple as  possible
through  'gnuplot_draw::draw_XXX'  static   methods.   This  could  be
changed in  the future with  a better OO approach  but for now,  it is
enough and seems to work.

Also some drawing functions' signatures have been modified to unify the
interface.