Ignore .ipp files in Doxygen
The various .ipp files in Bayeux, mainly used for Boost based 
serialization, cause many Doxygen warnings due to their odd inclusion
pattern and extensive use of macros.

Add .ipp to the list of files ignored by Doxygen. Whilst this removes
documentation for the serialization interface(s), these are not in
general for usage by clients. In addition, the warnings mean that we
probably shouldn't rely on the generated documentation.