Revisit the implementation of some 3D shapes
The  XXX::is_inside,  XXX::is_outside and  XXX::is_on_surface  virtual
methods are  reimplemented in such a  way they detect, within  a given
tolerance (the  skin thickness), if a  point is INSIDE, OUTSIDE  or on
one  of  the  FACES  of   a  solid.   The  former  implementation  was
considering that  the surface was part  of the INSIDE region  (in fact
the bound of the INSIDE region).

Impacted solid shapes are:

 - box
 - cylinder
 - tube
 - polycone
 - sphere
 - union 3D
 - intersection 3D
 - subtraction 3D

Not yet supported solids are:

 - polyhedron
 - tesselated solid

A default  version of  the i_shape_3d::is_outside  is provided  as the
negation  of the  XXX::is_inside  || XXX::is_on_surface(...  FACES_ALL
...)  expression for a given solid.


Also  the smart_id_locator  class and  the geom_map  class have  their
"locator"  implementation modified  to take  into account  the changes
explained above.  The  former implementation was detecting  if a point
was INSIDE a given solid using a given skin thickness as the tolerance
through  the i_shape_3d::is_inside  method.  Now  we determine  that a
point is  inside a solid  if the i_shape_3d::is_outside  method return
false. Of course,  using a skin thickness of  exactly zero corresponds
to tha case when we neglect the  ON_SURFACE case and their is only two
possibilities:  INSIDE  and  OUTSIDE.  It  may  occurs  that  in  some
circunstances, roundoff errors lead to a bad INSIDE/ON_SURFACE/OUTSIDE
classification even if some care has been taken to suppress this case.

The changes in  locator classes must be taken into  account in mctools
because  some algorithms  (step hit  backend processors)  need to  use
locators to compute the geometry Id  of the truth step hits during the
Geant4 particle tracking.