Deep refactorying of the implementation and minor change in the interface
I have  decided to  do this  important refactorying  because materials
(one of the elder modules in  Bayeux) was full of historical warts and
had   an  approximative   interface   as  well   as  a   questionnable
implementation in some parts.  It was not sane to release Bayeux 1.0.0
with this inside.  The interface is now simpler and the implementation
more robust (in principle). It should be easier to maintain and extend
in the future.

This  revision implies  slight changes  in  the interface  due to  the
removal of  direct access  to critical data  structures replaced  by a
proper public interface.  geomtools support for materials (GDML expor)
must then be adapted (see following commit).

This revision need to be validated by geomtools, mctools and Falaise
to make sure there is no side-effects, as expected.

This  commit  to:

- sanitize  the code,  replacing non-encapsulated  publicly accessible
  structs by classes with a safe public interface

- introduce a singleton table of elements loaded on first request

- introduce a  child class  isotope::id as the  unique id  for isotope
  objects (will be useful for other client apps/libs: genbb_help...)

- refactor access  to AME2003 isotope data  (mass.mas03 resource file)
  using a  singleton table of  isotopes (extensible for future  use if
  needed)

- fix a few  bugs in some resource files and  add more definitions for
  elements with natural isotopic abundance

- remove the JEFF 3.11 huge and  awkward data file (>40MB) and related
  code (within the materials::isotope class) for it is not used in the
  current implementation  (probably could  make sense within  a future
  release of genbb_help...  but wait and see).

- implement the support for  publishing arbitrary auxiliary properties
  associated to materials::material objects (possibly used by external
  application using  a materials::manager object). Makes  the embedded
  materials::factory  object more  discrete in  the materials::manager
  class (however this is not PIMPL).

- standardize the coding style and Doxygen the code

- fix the example