Add a new option"
XG faced issues while debugging some parts of Bayeux
because a lot of warning messages were interlaced with
some error messages of interest. In multi-threaded build mode
this is thus particularly tedious to exploit the messages.
The 'Bayeux_FORCE_CXX_ALL_WARNINGS' option is thus introduced
(ON by default) to enable the inhibition of -Wall compiler
option in such a case.