get back to previous way of setting momentum and particle energy
The G4ParticleGun interface is quite redundant in teh way to set
particle momentum and particle energy. Of course, setting one of the two
quantities fixes the other one i.e. setting momentum will set kinetic
energy and vice versa. The problem is that Geant4 dump message every
time one quantity changes the other. To avoid message flow, we get back
to the previous usage i.e. define the momentum direction and then the
kinetic energy of the particle. This suppose to get the particle mass so
this an additionnal operation but actually the calculation from Geant4
is the same (mass comes from PDG table).

The code has been also cleanup.