Add a combinaison of neutron physics
The mctools::g4::neutrons_physics_constructor  is added to  model some
physics  process  for  neutrons and  accompanying/related  process  of
interest.

A     possible     problem     has    been     detected     in     the
mctools::g4::sensitive_detector  class. The  track history  object was
cleared  from  the  EnfOfEvent  method  even  if  another  sensitive
detector instance needs it when its own EnfOfEvent  method is invoked.
The expected consequence is that such a
'secondary' sensitive detector can not access some specific properties
requested by  the user from  some sensitive categories  and associated
hits. The effect depends on the order of active sensitive detectors.
This is a bug. The solution consists in only clearing the track history object
(embedded in the Geant4 simulation manager) when we start a new event from the
event_action::BeginOfEventAction method.
The fix will be validated by Arnaud.

Cosmetics/style minor changes in a few files.