Fix attempt for malformed GDML syntax
The new support of "state", "temperature" and "pressure"
informations in materials through the GDML export function
was malformed.

This commit fixes the syntax for "state" while the way we should
generate the syntax for T and P is questionable. We have to investigate
both GDML requirements (XLS) and examples to find the right way to do it.
For now T and P GDML exports are removed.
XG+FM.