use vector_3d value in place of pointer
The local candidate vertex i.e. the one in the local frame coordinate
was initially a pointer to vector_3d object. The issue was that the
pointer's scope was related to _shoot_vertex function and thus the
adress of the pointer was destroyed before being used by any
vertex_validator. We change the local vertex candidate member into a
vector_3d and copying it each time we want to validate it (3 doubles to copy is not
a big deal).