new implementation for units support
The datatools/units.h and units.cc source files are redesigned with some
new  dedicated datatools::units::unit, datatools::units::unit_dimension
and datatools::units::registry classes.

The  datatools::units::unit class represents a unit in some unit system
with a name,  its symbol (and alternative ones if some exist) and its
intrinsic double value as provided by the CLHEP library.

The  datatools::units::unit_dimension class represents a category
of physical units like 'length', 'time', 'velocity' with its description
in term of dimensional analysis and the list of associated units.

The  datatools::units::registry is a smart container that allows to register
some sets of units and associated unit dimension and provide some methods
to fetch any information about units, symbol, dimensions. It is extensible
on the fly.

A system datatools::units::registry object is available as a static singleton.
It pre-registers a large set of SI and SI related units as well as a few
imperial and US units for convenience.

This new system is used for the parsing functionalities with
character strings like: "3.2 cm", "1.3 km/h", "1.117 MeV", "12 kpc".

This system has nothing to do with the Boost/Unit library but could be
interfaced to it in the future, if needed.