Add a new test program
Add  a  new   test  program  to  illustrate   the  deserialization  of
datatools::things  with old  class  export/serial tag  through a  mock
inherited class.  This usecase  (introduced for backward compatibility
of the deserialization of BiPo raw data file) implies a change in some
dedicated  serialization  macro (for  the  test1  test function):  the
BOOST_SERIALIZATION_BASIC_DECLARATION   macro  now   uses  "protected"
rather than  "private" for the  template serialize method in  a class.
However, we still have to investigate if this modification is relevant
because the technique used in the test2 test function maybe solves the
BiPo issue.