add a polycone model
The polycone model associates a geomtools::geom_id to a polycone shape
and then use the elementary polycone_vg generator to shoot vertices.

There is actually one limitation since the 'all_frustrum' mode must be
enabled : no selected frustrum of the polycone can be activated since
the polycone shape is known after the parsing of the properties
file. The number of frustrum is thus not known at the parsing time. A
better way to overpass this issue will be to send the setup file to the
initialize method of polycone_vg.