Fixes and improvements in the variant API, improvements of properties class
The variant_repository class is improved and the corresponding GUI
is updated.
Some fixes in the I/O tools for variants and properties.
The "examples/variants/" directory is added to illustrate some
of the functionalities of the datatools variants support.

The properties class now supports embedded unit for real number,
preserving not only the 'explicit unit' flag but also storing
the unit symbol parsed from a configuration file or set
programmatically ("mm", "kg"...).