Preliminary implementation of nuclear decay generation API
This  is   the  first  set  of   classes  used  by  the   new  generic
"genbb::nuclear_decay_generator"  class.   It  is based  on  the  formal
representation  of arbitrary  nuclear levels  associated to  arbitrary
decay table and related decay  processes (alpha and beta decays, gamma
transitions...).

In the  near future  we will  provide some  tools to  generate (semi-)
automatically some  sets of  configuration files describing  any decay
tables from  the ENSDF database  (Na22, Co60, Bi214...).   This should
make  the  maintenance  of  the  generators  of  interest  easier  and
will allow us to keep synchronized with up-to-date nuclear knowledge.

This is a preliminary commit, mainly to check the build within Bayeux.
This new piece of software should not be used for production because
it is not fully functionnal.