Add biasing weight information in base_step_hit
Preparing the implementation of  biasing techniques within Geant4, the
mctools::base_step_hit class  is now able  to record a  biasing weight
value. It is needed to keep this  information to be able to repair, at
the end of  a simulation run, the statistical bias  introduced by some
biasing algorithm (importance sampling).