Introduce a Qt interface class (ala G4Qt)
From within a given program, we cannot know in advance what (and when)
will initialize a QApplication object, needed for some GUI operations.

The "datatools::qt::interface" class (as done in Geant4) is introduced
to   allow   the  instantiation   or   the   reuse  of   an   existing
QApplication. We also  make sure that a well knwon  locale "C" is used
to  avoid issues  with LC_NUMERIC  (parsing with  fr_FR.XXX is  messy)
because QApplication  silently forces  some locales with  possbme side
effects.