Fix segfault at variant dialog destruction
Qt policy is rather difficult to understand when GUI objects
are to be destroyed. Here I refactored the instanciation of some instance
of the "datatools::configuration::ui::variant_repository_dialog"
class, not using a dynamically allocated object anymore because the
"delete" operator (and whatever technique I have tried) failed (segfault).