Fix a bug about the embedded library loader
The embedded library loader was destroyed just after its initialization
in the OCD driver object and the loading of requested libraries. However
its destruction was unloading the just loaded libraries, making them
unavailable for the rest of the run. OCD support was then unaccessible for
classes registered in such dynamically loaded libs.

This fix uses a scoped pointer to handle the embedded library loader in the
OCD driver object. The library loader is destroyed only at the end of the
driver life cycle, when the driver is destroyed.