Implement the biasing algo for primary particles
This algo enable to tag an event in such a way it should or not be tracked
by the simulation engine (i.e. Geant4).

It uses a combinaison of criteria to determine if an event
should be tracked or not, given the definition of a subset
of primary particles of interest and applying some selection
in terms of:
 - total kinetic energy
 - individual kinetic energy
 - direction of the momentum with respect to some
   collection of points of interest (PoI) which may be attractive
   (one keep only events with some particle pointing to some PoI)
   or repulsive (one reject events with some particle pointing to some PoI)

For now, only a 'master' particle is considered among a list of candidate
particles of interest and selection/rejection criteria are applied to it.

Example: This allow to express the following criteria:
 - keep only events with one gamma ray above 1 MeV pointing in the direction
   of the detector.
 - reject events with at least on electron with energy below 200 keV
   pointing in the direction of the veto detector.

The idea is to provide some tool to save CPU at simulations stage, avoiding
to track event of which the topology is guessed to be not efficient in term
of detector response (or dosimetry).



Future version will be able to apply selection/rejection criteria to
a set of particles of interest.