Enrich the primary_particle and i_genbb
It is now possible to assign an unique "generation id"
to each primary particle in a primary event. This new
attribute will enable in the future the association
between the primary particles and some Monte-Carlo hits
generated during the tracking of the event (i.e. track_id
from Geant4).

The genbb::primary_particle serialization now uses version 4.

The genbb::i_genbb base generator is given a new flag
to activate this functionality.