Fix the file location of BOOST_CLASS_VERSION macros
Formerly the BOOST_CLASS_VERSION macro was used from the
template implementation files for various XXX::serialize methods.
This macro sets explicitely the class version for the Boost/Serialization
library and archives.
The problem is that when one uses some I/O tool to store
a serializable object of versionned class XXX in some archive,
and this tool does not include the XXX.ipp file but only the XXX.h,
it cannot see the current version number of the
class and thus uses the default version number 0. This leads to some
load issue if the reader uses the proper version number.

I have decided to move the usage of BOOST_CLASS_VERSION in the XXX.h header
files for each serializable class XXX, in order to make sure any class
version number is known by any part of the code, regardless of the
inclusion or not of the XXX.ipp files.

This technique should be generalized to all versionned classes in Bayeux
and probably client library and software.