Add a flag in dpp::input_module
Add the 'clear_record' flag attributes in the dpp::input_module class.
This flag automatically makes the data record cleared
at each new load/process op.
If not set, this is the responsibility of the caller
to clear the data record from outside of the input module.
The input module 'process' method returns false if it finds
a non empty data record. This legacy behaviour was set in such a way
it is not possible to 'erase' a data record that has not explicitely
been cleared by the user.
Default value is false.