Add support for electric fields in G4 and sanitize the code
* A   preliminary   support   for   pure  electric   field   and   mixed
electromagnetic field is provided.

* The   more   general   class   mctools::g4::electromagnetic_field   is
implemented,      in     parallel      to     the      legacy     pure
mctools::g4::magnetic_field.    It  works   basically   in  the   same
way.  However, the  mctools::g4::detector_construction class  has been
refactored for the field support.  Also  note that Geant4 has a broken
field interface (bad signature for legacy method in G4ChordFinder) and
the mctools::g4::electromagnetic_field has been designed in such a way
it   inherits   the   G4MagneticField    class   in   place   of   the
G4ElectroMagneticField  class.   The   trick  breaks  the  inheritance
diagram as it should be but it is the only way to instantiate properly
a ChordFinder object that is needed to compute the curved track in the
presence of EM fields (by the way,  the G4 example itself uses a nasty
trick to make it work!).

This  is  still  preliminary  and  E-field support  will  need  to  be
validated (for example  by the LPCTrap simulation  group at Leuven/LPC
Caen).
I have  experienced some  spurious step error  while using  a (strong)
electric field with 1  MeV electrons in a simple setup.  I have made a
report to the G4 list and it seems that some bugs have been discovered
and fixed in more recent version  than the one we use (4.9.6.1).  This
has to be investigated.

* Some code have been improved (missing static const definitions...)

* Kozmetix.