Add a method in primary event class.
Add the genbb::primary_event::remove_particle method in order
to be able to remove some primary particles in a list.
This is a request from the Safe library which provide an AmBe
source decay generator that chains a Am-241 decay with the primary
alpha being captured by a Be-9 envelope, leading to the emission
of a neutron.