Publish the legacy C++ header from decay0
Up to now, the collection of C++ functions ported from the
Decay0/Genbb legacy package was private.

It is now published because it enables to build some new decay generators
that mix some of the decays supported by Decay0. This is a request from
the Safe client library for its AmBe generator.

Note: may be we could limit the public interface to generator functions
only and hide some Decay0 management headers that are only used through
the genbb::wdecay0 wrapper generator (to be investigated).