Refactor some components of the G4 module
Electromagnetic field support is revised in order to support an unique
class  for  the equation  of  motion  in  place  of the  four  classes
published in Geant4.

Support    for     region    is     refactor    through     the    new
mctools::g4::region_info and mctools::g4::regions_setup classes.

Support for electromagnetic extra models  by particle and by region is
introduced in order to treat differently ion in gas or solid materials
because  the default  ionization model  in  G4 is  not always  adapted
(on-the-fly assignment of an effective  charge to the dynamic particle
that biases the deviation of ion beam in an electric field).

Data output at  run/event action levels is revised and  can now use an
embedded dpp::output_module. Also PIMPL-ization of the I/O processing.

Add some options to the simulation manager and bxg4_production program.

Improvement  and update  of  the examples  (update  the config  files'
format,  the scripts).  Add  a  validation report  in  the ex02  about
charged ion deviation in some electric field.

Kozmetix.