Finalize the emfield::multi_zone_field class
The    emfield::multi_zone_field   class    enables   to    model   an
electromagnetic field which is known/computable in different (possibly
contiguous) *zones*.   A zone is represented  by a shape with  its own
positioning in a given  geometry setup.  The corresponding geometrical
volume is  associated to some  EM field. Note  that the volume  is not
related  to any  logical/physical  volume available  from a  geomtools
geometry manager. This is because we have a priori no relation between
the modelling  of a geometry  virtual setup through geomtools  and the
constraints at work while computing a field map with PDE solver.

A special policy  is used when a point lies  in the overlapping region
between 2 zones.  We have introduced  a zone *priority* that tells the
field computing driver if we should use the field value of one zone or
the other  in the overlapping volume.   In case of same  priority, the
mean value  of the  field is used.  By default, a  zone is  assigned a
minimum priority.

This makes possible to use zones  with a high precision field of which
the field  value will always  be computed and  used, and use  only low
precision field values outside this zone.

The class now uses a  geomtools::shape_factory that can be passed from
a geometry service  or embedded. A multi_zone_field can  thus be fully
setup from a configuration file.

Example  ex01   is  updated   and  now  illustrates   the  use   of  a
multi_zone_field object.