First import of the LaHague Bayeux module source directory
LaHague is a new module for Bayeux dedicated to dosimetry and
radioprotection. It aims to provide dedicated backends algorithms
for dose computation from Monte-Carlo output from Bayeux/mctools (Geant4).

This is a very preliminary version adapted from a preliminary work
done by Arnaud Chapon in the framework of his radioprotection activities
for the Cerap company and LPC Caen.