Extend the vertex generator interface to randomized time
For a certain category of physics study, the generation
of random vertex within some geometry is correlated to
some absolute time reference. Typically, the vertex PDF
depends on time. This is the case for example for migrating
unstable isotopes or decaying particle trapped within some
complex time dependant field.

The current commit enables to set a vertex generator
as a vertex/time generator. New methods are provided to check
if  a generator can also generate the (decay) time associated
to a randomized vertex.

This modifications does not change the existing interface of
previous generators.

The programmer is requested to use the
"genvtx::i_vertex_generator::_set_time_generator" method
from the constructor of a new generator class to indicate
if this generator should also generate decay time.

The virtual method:

  void _shoot_vertex(mygsl::rng & random_, geomtools::vector_3d & vertex_);

must be provided if no time is generated.

The new virtual method

  void _shoot_vertex_and_time(mygsl::rng & random_,
                              geomtools::vector_3d & vertex_,
                              double & time_);

must be provided if the time is generated.

The base "genvtx::i_vertex_generator" provides default implementation that throws.

The "genvtx::genvtx_driver", "genvtx::manager" and "genvtx::from_file_vg" classes are adapted
to this new feature.