Fix reference time issue in genbbsub
We now  force the decay  time of  so-called "background" events  to be
zero. Formerly it was randomized  w.r.t. the half-time of the decaying
isotope. A  consequence was that  some very very long  reference times
were  set  for  some  isotopes  (Tl208: ~1e11  s)  and  produced  some
arbitrary reference time with not significant gamma emission times due
to roundoff errors.