Add tools for diagnostic
Some users have met some problems concerning the automatic detection
of the library installation prefix through the BinReloc tool
on Ubuntu 15.04/ Despite the system is detected with multiarch support
by CMake, and properly installed within the "lib/x86_64-linux-gnu/"
subdirectory. The "br_find_prefix" function failed to detect it dynamically
and returns the wrong parent path. The resolution of the Bayeux prefix directory
is thus broken. The loading of published resource files fails.
This is a preliminary fix proposal. This will be reviewed later
when I will have such a system at end to reproduce the error.
This fix should not have impact on existing installation (I hope so!).