Add new definition basic materials
The original isotope/element/material definition files
should now be considered as obsolete.
We prefer to rely on a versioned list of *basic* materials
and associated aliases named after the pattern ``basic::XXXXX``.
This makes things easier to maintain and compare.
Also ``std::`` prefix was confusing. It is only maintained here for
backward compatibility.