Refactor some parts of the build process
The CMake build  scripts are modified in  such a way we  can now build
Bayeux with only some of the library modules. Of course  Bayeux/datatools
remains the mandatory core module.
The other packages  (cuts, brio... lahague) as well
as extension  library modules (Geant4/MCNP)  are optional.  This  is a
request from new Bayeux users who want  to be able to build and use Bayeux
only with the minimal set of features they need (for example up to geomtools).
Most packages are enabled by default but experimental ones.

MCNP  is very  very  prelimary. It  consists on  some  initial set  of
classes to build an interface for MCNP from geomtools and mctools.  It
is disabled by default. It is one of the main new task to do in Bayeux
beside SuperNEMO activities, for another project/contract we have at
LPC Caen.


New available options :

 * Bayeux_WITH_BRIO
 * Bayeux_WITH_CUTS
 * Bayeux_WITH_MYGSL
 * Bayeux_WITH_DPP
 * Bayeux_WITH_MATERIALS
 * Bayeux_WITH_GEOMTOOLS
 * Bayeux_WITH_EMFIELD
 * Bayeux_WITH_GENBB_HELP
 * Bayeux_WITH_GENVTX
 * Bayeux_WITH_MCTOOLS
 * Bayeux_WITH_LAHAGUE
 * Bayeux_BUILD_MCNP_MODULE

Preserved options:

 * Bayeux_BUILD_GEANT4_MODULE

The CMake script for Bayeux/mctools has been refactored to ease
maintenance.

Also a new option Bayeux_USE_LEGACY_ROOT is introduced to enable
the use of a legacy version of ROOT (v5.3X) from a previous version of Cadfael.
This is transitional. This is because my own CadfaelBrew breaks at build
and I still cannot make it work flawlessly. As soon this issue will be
solved (for me and possibly others), we will be able to remove this option
and run the new ROOT tool support.

Updated README.rst.

Kozmetix.

This commit attempts to close a set of unstable revisions that fixes several
issues reported by Ben and also include some new stuff that I have been
working on for a few weeks. It builds and passes all tests with gcc 4.8.4 and
clang 3.4 on my Xubuntu 14.04 box. However, as usual, we should check it,
particularly with Falaise.