Add support for include-what-you-use
New build option BAYEUX_WITH_IWYU_CHECK, OFF by default.

When set to ON, searches for the include-what-you-use tool 

http://include-what-you-use.org

and if found, sets the CXX_INCLUDE_WHAT_YOU_USE target property
on Bayeux to the executable. Diagnostics will then be emitted when
building on any #includes that can be removed or need adding.

NOTE: Requires that include-what-you-use has access to clang
builtin headers. See 

https://github.com/include-what-you-use/include-what-you-use/issues/100

for details.